+++ data/src/main/java/com/fernandocejas/android10/sample/data/entity/weather/Sys.java
@@ -3,12 +3,10 @@ package com.fernandocejas.android10.sample.data.entity.weather;
 
 
 public class Sys {
-
-    public Integer type;
-    public Integer id;
-    public Float message;
+    public float type;
+    public float id;
+    public float message;
     public String country;
-    public Integer sunrise;
-    public Integer sunset;
-
+    public float sunrise;
+    public float sunset;
 }
