+++ data/src/main/java/com/fernandocejas/android10/sample/data/entity/weather/Wind.java
@@ -1,10 +1,6 @@
-
 package com.fernandocejas.android10.sample.data.entity.weather;
 
-
 public class Wind {
-
-    public Float speed;
-    public Integer deg;
-
+  public float speed;
+  public float deg;
 }
